requirements.txt
torch>=2.0.0
deepxde>=1.9.3
numpy>=1.24.3
pandas>=2.0.0
pyyaml>=6.0
scipy>=1.10.0
matplotlib>=3.7.1
ensure
python-box
# yaml is provided by pyyaml above